requirements.txt
flet==0.26.0
googlesearch_python==1.3.0
Requests==2.32.3
